Changed state system to use C++11 enum classes
Changes to state system:
+ Changed state mechanism to use an C++11 `enum class`.
+ Made `State` `enum` use C++11 size definition technique.
+ Changed state variable to `State` type.
+ Changed state variable assignment from `0` to `State::Title`.
+ Refactored all uses of `state_title`, `state_menu`, `state_game` and `state_pause`. 

For information on 'enum class` and defining an enum's size, see http://www.cprogramming.com/c++11/c++11-nullptr-strongly-typed-enum-class.html.
I'm unsure if `uint8_t` is available here, but if it is it would be preferable to `char`. `byte` would also be preferable, but I believe `byte` is just a `typedef`d `std::uint8_t`.